Refactors generators
- split generators by separate files
- add extra test case to check RAML doc header
